everest-server/pyproject.toml
[project]
name = "everest-server"
version = "0.1.0"
description = "Background service for the Everest app."
authors = [
    {name = "pugsedo"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
packages = [{ include = "everest_server", from = "src" }]
dependencies = [
    "llama-cpp-python (>=0.3.16,<0.4.0)",
    "fastapi (>=0.121.0,<0.122.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "pyautogui (>=0.9.54,<0.10.0)",
]

[dependency-groups]
dev = [
    "pyinstaller (>=6.16.0,<7.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
everest = "everest_server.app:run_app"
